main: Refactor enemy.h/.cpp, refactor entity movement and animation for reusability (dry), remove size_t where it doesn't make sense.

+++ include/game_clock.h
@@ -5,12 +5,14 @@
 
 #include <SFML/System/Clock.hpp>
 
+#include <cstdint>
+
 class GameClock
 {
     public:
         // World Clock
         float getWorldTimeSeconds();
-        int getWorldTimeMs();
+        uint64_t getWorldTimeMs();
 
         // Delta Clock
         void restartDeltaClock();
